Finish adding in errors
